CyberGuard Database-Driven Authentication and Authorization System
"Simple Local Server with Database Integration (WAMP Server), enabling the deployment of web pages and applications on a personal computer without the need for external hosting.

The developed website primarily consists of HTML, CSS, JavaScript, and jQuery files, presenting visual content and interactive elements in a web browser. Configured relationships with the database dictate the visibility permissions of specific pages. Depending on the permissions granted in the database, the page either displays or denies access (in case of authorization refusal). Passwords in the database must be configured for MD5 encryption.

The 'Insufficient Permissions' page serves as a notification for inadequate access rights. To run the website, one must install a local server, configure it, and place the website files in the appropriate directory. After launching the server, the website can be viewed on the computer via http://localhost/."